# [Jiterpreter] Add support for TryGetHashCode intrinsic