@AustinWise 2023-02-04
@kg is there a documented way to test and benchmark Jiterpreter changes? I had [BenchmarkDotNet project](https://github.com/AustinWise/TryGetHashCodeBenchmark/tree/wasm) that defined some custom WASM jobs, but it appears that no longer works.