@kg 2023-02-04
> @kg is there a documented way to test and benchmark Jiterpreter changes? I had [BenchmarkDotNet project](https://github.com/AustinWise/TryGetHashCodeBenchmark/tree/wasm) that defined some custom WASM jobs, but it appears that no longer works.

It should work the same as anything else, you can manually adjust the tiering thresholds for it so that traces compile sooner (either by passing a runtime arg, or editing the defaults in [options-def.h](https://github.com/dotnet/runtime/blob/f8daacff9d5ac3cad277fba8a97aefe72865ae45/src/mono/mono/utils/options-def.h#L113)). 

To set a runtime arg you can [pass it from JS when initializing the runtime](https://github.com/dotnet/runtime/blob/f8daacff9d5ac3cad277fba8a97aefe72865ae45/src/mono/sample/wasm/simple-raytracer/main.js#L23).