@AustinWise 2023-02-05
I found the incantation to make [my benchmark](https://github.com/AustinWise/TryGetHashCodeBenchmark/tree/wasm) work again:

```bash
./dotnet.sh build -p:TargetOS=browser -p:TargetArchitecture=wasm -c Release src/mono/wasm/Wasm.Build.Tests /t:InstallWorkloadUsingArtifacts
```

This creates the `dotnet-latest` directory containing a .NET SDK that can be used with BenchmarkDotNet.

Anyways, the results of running the benchmark:

``` ini
BenchmarkDotNet=v0.13.3, OS=ubuntu 22.04
AMD Ryzen Threadripper PRO 3955WX 16-Cores, 1 CPU, 32 logical and 16 physical cores
.NET SDK=7.0.102
  [Host]     : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  PR         : .NET Core (Mono) 8.0.0-dev, Wasm AOT
  merge-base : .NET Core (Mono) 8.0.0-dev, Wasm AOT

Runtime=Wasm  IterationCount=15  LaunchCount=2  
WarmupCount=10  
V8 version 11.1.92
```

|                 Method |        Job |        Toolchain | NumberOfObjects |     Mean |   Error |  StdDev | Ratio | RatioSD |
|----------------------- |----------- |----------------- |---------------- |---------:|--------:|--------:|------:|--------:|
| TryGetNonExistentValue |         PR |         Wasm: PR |            1000 | 132.9 ns | 0.83 ns | 1.21 ns |  1.01 |    0.01 |
| TryGetNonExistentValue | merge-base | Wasm: merge-base |            1000 | 131.0 ns | 1.07 ns | 1.56 ns |  1.00 |    0.00 |
|                        |            |                  |                 |          |         |         |       |         |
|    TryGetExistingValue |         PR |         Wasm: PR |            1000 | 198.4 ns | 2.41 ns | 3.61 ns |  0.90 |    0.02 |
|    TryGetExistingValue | merge-base | Wasm: merge-base |            1000 | 219.8 ns | 0.96 ns | 1.41 ns |  1.00 |    0.00 |