@kg 2023-02-06
You'll need to rebase this now to resolve a conflict (sorry! 🙇‍♀️), it shouldn't be too hard to do though. The jiterpreter code just got rearranged and split into two files, but copying your work over should be a simple bit of copy-paste.