@kg 2023-02-07
> I resolved the merge conflicts and the tests pass.

Great! I'll make sure we're cool adding this (I don't see any problems) and get it merged.